requirements.txt
requests
openai
opencv-python
numpy
pillow
